+++ envoy_build_config/extension_registry.cc
@@ -6,12 +6,13 @@
 #include "source/extensions/clusters/dynamic_forward_proxy/cluster.h"
 #include "source/extensions/compression/brotli/decompressor/config.h"
 #include "source/extensions/compression/gzip/decompressor/config.h"
+#include "source/extensions/filters/http/alternate_protocols_cache/config.h"
 #include "source/extensions/filters/http/buffer/config.h"
 #include "source/extensions/filters/http/decompressor/config.h"
 #include "source/extensions/filters/http/dynamic_forward_proxy/config.h"
 #include "source/extensions/filters/http/router/config.h"
 #include "source/extensions/filters/network/http_connection_manager/config.h"
-#include "source/extensions/http/header_formatters/preserve_case/preserve_case_formatter.h"
+#include "source/extensions/http/header_formatters/preserve_case/config.h"
 #include "source/extensions/http/original_ip_detection/xff/config.h"
 #include "source/extensions/stat_sinks/metrics_service/config.h"
 #include "source/extensions/transport_sockets/raw_buffer/config.h"
@@ -35,6 +36,10 @@ void ExtensionRegistry::registerFactories() {
       forceRegisterBrotliDecompressorLibraryFactory();
   Envoy::Extensions::Compression::Gzip::Decompressor::forceRegisterGzipDecompressorLibraryFactory();
   Envoy::Extensions::Http::OriginalIPDetection::Xff::forceRegisterXffIPDetectionFactory();
+  Envoy::Extensions::Http::HeaderFormatters::PreserveCase::
+      forceRegisterPreserveCaseFormatterFactoryConfig();
+  Envoy::Extensions::HttpFilters::AlternateProtocolsCache::
+      forceRegisterAlternateProtocolsCacheFilterFactory();
   Envoy::Extensions::HttpFilters::Assertion::forceRegisterAssertionFilterFactory();
   Envoy::Extensions::HttpFilters::Decompressor::forceRegisterDecompressorFilterFactory();
   Envoy::Extensions::HttpFilters::BufferFilter::forceRegisterBufferFilterFactory();